+++ web/src/features/weather-layers/wind-layer/WindLayer.tsx
@@ -10,7 +10,6 @@ import {
   windLayerLoadingAtom,
 } from 'utils/state/atoms';
 import { useReferenceWidthHeightObserver } from 'utils/viewport';
-
 import { Windy } from './windy';
 
 let windySingleton: Windy | null = null;
