+++ web/src/features/weather-layers/wind-layer/windy.test.ts
@@ -1,6 +1,7 @@
 import { round } from 'utils/helpers';
 
-import { bilinearInterpolateVector, windIntensityColorScale } from './windy';
+import { bilinearInterpolateVector } from './calc';
+import { windIntensityColorScale } from './scales';
 
 describe('windIntensityColorScale', () => {
   it('should return an array of colors', () => {
